% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{change_brightness}
\alias{change_brightness}
\title{Change brightness value of a hex color.}
\usage{
change_brightness(hex_col, val_fact = 0.25)
}
\arguments{
\item{hex_col}{hex color}

\item{val_fact}{brightness percentage change}
}
\description{
Function to change brightness value of a hex color.
Using this will create colors not from the official palette.
}
\examples{
# make brighter
den_palettes$secondarydark |> change_brightness(-.5) |> color_bars()
# make darker
den_palettes$secondarydark |> change_brightness(-.5) |> color_bars()

}
