% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{color_bars}
\alias{color_bars}
\title{Convenience demo function}
\usage{
color_bars(col = den_colors)
}
\arguments{
\item{col}{color hex values}
}
\description{
Display color bars for submitted hex codes
}
\examples{
color_bars(den_colors)
color_bars(den_colors[c("red","neutralblue")])
color_bars(den_palettes$secondarydark)

}
