% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{den_colors}
\alias{den_colors}
\title{Set of hex values of official Denison colors}
\format{
A named character vector of hex values
}
\source{
<https://denison.edu/offices/communications/branding/color-palette>
}
\usage{
den_colors
}
\description{
Set of hex values of official Denison colors
}
\keyword{datasets}
