% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{den_colors_df}
\alias{den_colors_df}
\title{Dataframe of official Denison palettes, their color names and naming codes}
\format{
A data frame of official Denison colors with hex, CMYK and
Pantone codes
\describe{
   \item{group}{palette group}
   \item{name}{color name}
   \item{hex}{hex values of colors in that palette}
   \item{cmyk}{CMYK code}
   \item{pms}{Pantone code}
}
}
\source{
<https://denison.edu/offices/communications/branding/color-palette>
}
\usage{
den_colors_df
}
\description{
Dataframe of official Denison palettes, their color names and naming codes
}
\keyword{datasets}
