% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{den_cols}
\alias{den_cols}
\title{Extract den_colors as hex codes}
\usage{
den_cols(...)
}
\arguments{
\item{...}{A Vector of character names of den_colors}
}
\value{
A vector of hex codes
}
\description{
Convert names of den_colors to hex codes
}
\examples{
den_cols("tasseldarkgold")

}
