% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{den_font_map}
\alias{den_font_map}
\title{Names of open source fonts that are similar to the official Denison fonts.}
\format{
A data frame of font names
\describe{
   \item{purpose}{place to use font}
   \item{denison_font}{Offical font, usually closed source Adobe}
   \item{denison_alt}{Sanctioned Alternative}
   \item{google_substitute}{Open source Google font used in this package}
}
}
\source{
<https://denison.edu/offices/communications/branding/typography>
}
\usage{
den_font_map
}
\description{
These are open source fonts from Google Fonts that are similar to
the official Denison fonts.
}
\keyword{datasets}
