% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{den_fonts}
\alias{den_fonts}
\title{Google Substitute Names of Denison  fonts}
\format{
A named character vector of font names
}
\source{
<https://denison.edu/offices/communications/branding/color-palette>
}
\usage{
den_fonts
}
\description{
Set of  Google Substitute Names and aliases of Denison  fonts
}
\keyword{datasets}
