% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{den_pal}
\alias{den_pal}
\title{Interpolate colors of a `den` color palette}
\usage{
den_pal(palette = "primary", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in den_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\value{
Returns a function that takes an integer argument
(the required number of colors) and returns a character vector of colors
}
\description{
Function to interpolate a den color palette
Mainly for internal use.
Using this will create colors not from the official palette.
}
