% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{den_palettes}
\alias{den_palettes}
\title{List of official Denison palettes and their colors}
\format{
A grouped list of official Denison palettes with hex values  of the colors
}
\source{
<https://denison.edu/offices/communications/branding/color-palette>
}
\usage{
den_palettes
}
\description{
List of official Denison palettes and their colors
}
\keyword{datasets}
