% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denisonbrand-package.R
\docType{package}
\name{denisonbrand}
\alias{denisonbrand-package}
\alias{denisonbrand}
\title{Denison Brand styles}
\description{
Typefaces and Color palettes for ggplot2 using Denison University Branding as of 2023.
}
\section{Credit}{

Branding guildlines at
\url{https://brand.denison.edu}
Color tools Modeled on
\url{https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/apsteinmetz/denisonbrand}
  \item Report bugs at \url{https://github.com/apsteinmetz/denisonbrand/issues}
}

}
\author{
Arthur Steinmetz
}
\keyword{internal}
