% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_functions.R
\name{ggden_line}
\alias{ggden_line}
\title{Den-style geom for line or dot plots}
\usage{
ggden_line(pal = "secondarydark")
}
\arguments{
\item{pal}{override default den secondarydark palette}
}
\description{
Use to set line style and theme with one call.  Meta ggplot parameter for line or dot
plots that sets all den-relevant style
elements to defaults, including theme.
}
\examples{
library(ggplot2)
library(dplyr)
library(showtext)
load_fonts()
# plot ranking over time for Denison, Colgate and Lafayette
rank_data |>
 filter(college \%in\% c("Denison University","Colgate University","Lafayette College")) |>
  filter(year > 2002) |>
  ggplot(aes(x = year, y = rank, color = college)) +
  ggden_line() +
  scale_y_reverse() +
  labs(title = "U.S. News & World Report Liberal Arts College Rankings",
       subtitle = "Denison, Colgate and Lafayette",
       caption = 'Source: Andrew G. Reiter, “U.S. News & World Report
       Historical Liberal Arts College and University Rankings,”
       \navailable at: http://andyreiter.com/datasets/',
       x = "Year",
       y = "Rank (lower is better)",
       color = "College")

}
