% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_functions.R
\name{load_fonts}
\alias{load_fonts}
\title{Load Google Fonts}
\format{
A named character vector of font names ready for use with showtext
}
\source{
<https://denison.edu/offices/communications/branding/color-palette>
}
\usage{
load_fonts()
}
\description{
Load Google versions of Denison Fonts into Environment
}
\examples{
library(ggplot2)
library(showtext)
showtext_auto()

load_fonts()
gg <- ggplot(NULL, aes(x = 1, y = 1)) + ylim(0.5, 1.5) + theme_void()
for (n in 1:length(den_fonts)){
gg <- gg + annotate("text", 1, 1.4 - (0.1 * n),
                     family = den_fonts[n], size = 15,
                     label = paste(names(den_fonts[n]),": ",den_fonts[n]))
}
gg

showtext_auto(enable = FALSE)
}
