% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{peer_schools_list}
\alias{peer_schools_list}
\title{List of Denison peer schools}
\format{
List chr [1:20]
}
\usage{
peer_schools_list
}
\description{
List of Denison peer schools
}
\keyword{datasets}
