% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{rank_data}
\alias{rank_data}
\title{Top 100 Liberal Arts College and University Rankings}
\format{
A data frame
\describe{
   \item{college}{College Name}
   \item{ipeds_id}{IPEDS ID}
   \item{state}{State}
   \item{year}{Year}
   \item{rank}{Rank}
}
}
\source{
<http://andyreiter.com/datasets/>
}
\usage{
rank_data
}
\description{
U.S. News & World Report Historical Liberal Arts College and University Rankings.
A subset of the data from Andy Reiter's College Scorecard Data.
Available at: http://andyreiter.com/datasets/
}
\keyword{datasets}
