% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_fill_den}
\alias{scale_fill_den}
\title{Fill scale constructor for den colors in ggplot2}
\usage{
scale_fill_den(
  palette = "secondarydark",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in den_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for den colors in ggplot2
}
\details{
The default palette is "secondarydark" which produces sensible results in most cases.  Use
scale_color_manual(values=den_cols(<color name(s)>) for finer control.
}
\examples{
library(ggplot2)
library(dplyr)
library(showtext)
load_fonts()
pell_grants |>
  filter(college \%in\%
    c("Denison University","Colgate University","Lafayette College","Bucknell University")) |>
  ggplot(aes(college,pell_frac,fill = college)) +
  geom_col() +
  labs(title = "U.S. News & World Report\nLiberal Arts College Rankings",
       subtitle = "Denison, Bucknell, Colgate and Lafayette",
       caption = 'Source: U.S. News & World Report, available at:
       https://www.usnews.com/best-colleges/rankings/
       national-liberal-arts-colleges/economic-diversity',
       y = "Fraction of Students Receiving Pell Grants") +
  scale_fill_den(palette = "secondarydark") +
  theme_den()

}
