% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_functions.R
\name{theme_den}
\alias{theme_den}
\title{Denison Brand Theme for non-data elements of a plot}
\usage{
theme_den(...)
}
\arguments{
\item{...}{override or supplement arguments to pass through to theme()}
}
\description{
Denison Brand Theme for non-data elements of a plot
}
