% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_functions.R
\name{ggden}
\alias{ggden}
\title{Den-style elements for filled plots}
\usage{
ggden(pal = "secondarydark")
}
\arguments{
\item{pal}{override default den secondarydark palette}
}
\description{
Meta ggplot parameter for filled plots that sets all den-relevant style
elements to defaults, including theme.  Still needs a geom.
}
\examples{
library(ggplot2)
library(showtext)
load_fonts()
dt = data.frame(Category = LETTERS[1:6], values = runif(6)*10)
dt |> ggplot(aes(Category, values,fill = Category)) + geom_col() +
  ggden() +
  labs(title = "Denison Brand Colors")
}
