% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_functions.R
\name{ggden_line}
\alias{ggden_line}
\title{Den-style geom for line or dot plots}
\usage{
ggden_line(pal = "secondarydark")
}
\arguments{
\item{pal}{override default den secondarydark palette}
}
\description{
Meta ggplot parameter for line or dot plots that sets all den-relevant style
elements to defaults
}
\examples{
library(ggplot2)
# Create a sequence of years
years <- seq(2014, 2023)

# Create series
series <- LETTERS[1:6]

# Create data frame
df <- data.frame(year = rep(years, each = length(series)),
                 series = rep(series, times = length(years)),
                 value = (runif(length(years) * length(series))+1)*1:6)

# plot df using ggplot
df |> ggplot(aes(year, value, color = series)) +
  ggden_line()

}
