% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_color_den}
\alias{scale_color_den}
\title{Color scale constructor for den colors in ggplot2}
\usage{
scale_color_den(
  palette = "secondarydark",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in den_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for den colors in ggplot2
}
\details{
The default palette is "secondarydark" which produces sensible results in most cases.  Use
scale_color_manual(values=den_cols(<color name(s)>) for finer control.
}
\examples{
library(ggplot2)
library(dplyr)
library(showtext)
load_fonts()
rank_data |>
 filter(college \%in\% c("Denison University","Colgate University","Lafayette College")) |>
  filter(year > 2002) |>
  ggplot(aes(x = year, y = rank, color = college)) +
  geom_line(linewidth = 1) +
  scale_y_reverse() +
  labs(title = "U.S. News & World Report\nLiberal Arts College Rankings",
       subtitle = "Denison, Colgate and Lafayette",
       caption = 'Source: Andrew G. Reiter, “U.S. News & World Report
       Historical Liberal Arts College and University Rankings,”
       \navailable at: http://andyreiter.com/datasets/',
       x = "Year",
       y = "Rank (lower is better)",
       color = "College") +
  scale_color_den(palette = "secondarydark") +
  theme_den()

}
