% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_color_den}
\alias{scale_color_den}
\title{Color scale constructor for den colors in ggplot2}
\usage{
scale_color_den(
  palette = "secondarydark",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in den_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for den colors in ggplot2
}
\details{
The default palette is "secondarydark" which produces sensible results in most cases.  Use
scale_color_manual(values=den_cols(<color name(s)>) for finer control.
}
\examples{
library(ggplot2)
# Create a sequence of years
years <- seq(2014, 2023)

# Create series
series <- LETTERS[1:6]

# Create data frame
df <- data.frame(year = rep(years, each = length(series)),
                 series = rep(series, times = length(years)),
                 value = (runif(length(years) * length(series))+1)*1:6)

df |> ggplot(aes(year, value, color = series)) +
  geom_line(linewidth = 1) +
  scale_color_den(palette = "secondarydark") +
  theme_den()

}
