% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_fill_den}
\alias{scale_fill_den}
\title{Fill scale constructor for den colors in ggplot2}
\usage{
scale_fill_den(
  palette = "secondarydark",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in den_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for den colors in ggplot2
}
\details{
The default palette is "secondarydark" which produces sensible results in most cases.  Use
scale_color_manual(values=den_cols(<color name(s)>) for finer control.
}
\examples{
library(ggplot2)
dt = data.frame(Category = LETTERS[1:6], values = runif(6)*10)
dt |> ggplot(aes(Category, values,fill = Category)) + geom_col() +
scale_fill_den() +
theme_den()

}
